// web/src/types/interfaces.ts
export interface IComfyUserSettings {
    storageLocation: string | null;
    multiUserServer: boolean | null;
    isNewUserSession: boolean | null;
}